Move NPC dialogue functions into separate file
