+++ src/Engine/Objects/NPC.cpp
@@ -106,90 +106,6 @@ bool CheckHiredNPCSpeciality(NPCProf prof) {
         || pParty->pHirelings[1].profession == prof;
 }
 
-//----- (004B40E6) --------------------------------------------------------
-void NPCHireableDialogPrepare() {
-    int v0 = 0;
-    NPCData *v1 = houseNpcs[currentHouseNpc].npc;
-
-    pDialogueWindow->Release();
-    pDialogueWindow = new GUIWindow(WINDOW_Dialogue, {0, 0}, {render->GetRenderDimensions().w, 350}, 0);
-    pBtn_ExitCancel = pDialogueWindow->CreateButton({471, 445}, {169, 35}, 1, 0,
-        UIMSG_Escape, 0, Io::InputAction::Invalid, localization->GetString(LSTR_CANCEL), {ui_exit_cancel_button_background}
-    );
-    pDialogueWindow->CreateButton({0, 0}, {0, 0}, 1, 0, UIMSG_HouseScreenClick, 0);
-    if (!pNPCStats->pProfessions[v1->profession].pBenefits.empty()) {
-        pDialogueWindow->CreateButton({480, 160}, {140, 30}, 1, 0,
-            UIMSG_ClickNPCTopic, DIALOGUE_PROFESSION_DETAILS, Io::InputAction::Invalid, localization->GetString(LSTR_MORE_INFORMATION)
-        );
-        v0 = 1;
-    }
-    pDialogueWindow->CreateButton({480, 30 * v0 + 160}, {140, 30}, 1, 0,
-        UIMSG_ClickNPCTopic, DIALOGUE_HIRE_FIRE, Io::InputAction::Invalid, localization->GetString(LSTR_HIRE));
-    pDialogueWindow->_41D08F_set_keyboard_control_group(v0 + 1, 1, 0, 2);
-    dialog_menu_id = DIALOGUE_OTHER;
-}
-
-void updateNPCTopics(int npc) {
-    int num_menu_buttons = 0;
-
-    currentHouseNpc = npc;
-    Sizei renDims = render->GetRenderDimensions();
-    if (houseNpcs[npc].type == HOUSE_TRANSITION) {
-        pDialogueWindow->Release();
-        pDialogueWindow = new GUIWindow(WINDOW_Dialogue, {0, 0}, renDims, 0);
-        transition_button_label = houseNpcs[npc].label;
-        pBtn_ExitCancel = pDialogueWindow->CreateButton({566, 445}, {75, 33}, 1, 0, UIMSG_Escape, 0, Io::InputAction::No, localization->GetString(LSTR_CANCEL), {ui_buttdesc2});
-        pBtn_YES = pDialogueWindow->CreateButton({486, 445}, {75, 33}, 1, 0, UIMSG_HouseTransitionConfirmation, 1, Io::InputAction::Yes, transition_button_label, {ui_buttyes2});
-        pDialogueWindow->CreateButton({pNPCPortraits_x[0][0], pNPCPortraits_y[0][0]}, {63, 73}, 1, 0, UIMSG_HouseTransitionConfirmation, 1,
-                                      Io::InputAction::EventTrigger, transition_button_label);
-        pDialogueWindow->CreateButton({8, 8}, {460, 344}, 1, 0, UIMSG_HouseTransitionConfirmation, 1, Io::InputAction::Yes, transition_button_label);
-    } else {
-        if (dialog_menu_id == DIALOGUE_OTHER) {
-            pDialogueWindow->Release();
-        } else {
-            for (int i = 0; i < houseNpcs.size(); ++i) {
-                houseNpcs[i].button->Release();
-                houseNpcs[i].button = nullptr;
-            }
-        }
-        pDialogueWindow = new GUIWindow(WINDOW_Dialogue, {0, 0}, {renDims.w, 345}, 0);
-        pBtn_ExitCancel = pDialogueWindow->CreateButton({471, 445}, {169, 35}, 1, 0, UIMSG_Escape, 0, Io::InputAction::Invalid,
-                                                        localization->GetString(LSTR_END_CONVERSATION), {ui_exit_cancel_button_background});
-        pDialogueWindow->CreateButton({8, 8}, {450, 320}, 1, 0, UIMSG_HouseScreenClick, 0);
-        dialog_menu_id = DIALOGUE_MAIN;
-        if (houseNpcs[npc].type == HOUSE_PROPRIETOR) {
-            window_SpeakInHouse->initializeDialog();
-        } else { // NPC
-            NPCData *npcData = houseNpcs[npc].npc;
-            if (npcData->is_joinable) {
-                num_menu_buttons = 1;
-                pDialogueWindow->CreateButton({480, 160}, {140, 30}, 1, 0, UIMSG_ClickNPCTopic, DIALOGUE_13_hiring_related);
-            }
-
-            #define AddScriptedDialogueLine(EVENT_ID, MSG_PARAM) \
-                if (EVENT_ID) { \
-                    if (num_menu_buttons < 4) { \
-                        int res = npcDialogueEventProcessor(EVENT_ID); \
-                        if (res == 1 || res == 2) \
-                            pDialogueWindow->CreateButton({480, 160 + 30 * num_menu_buttons++}, {140, 30}, 1, 0, UIMSG_ClickNPCTopic, MSG_PARAM); \
-                    } \
-                }
-
-            AddScriptedDialogueLine(npcData->dialogue_1_evt_id, DIALOGUE_SCRIPTED_LINE_1);
-            AddScriptedDialogueLine(npcData->dialogue_2_evt_id, DIALOGUE_SCRIPTED_LINE_2);
-            AddScriptedDialogueLine(npcData->dialogue_3_evt_id, DIALOGUE_SCRIPTED_LINE_3);
-            AddScriptedDialogueLine(npcData->dialogue_4_evt_id, DIALOGUE_SCRIPTED_LINE_4);
-            AddScriptedDialogueLine(npcData->dialogue_5_evt_id, DIALOGUE_SCRIPTED_LINE_5);
-            AddScriptedDialogueLine(npcData->dialogue_6_evt_id, DIALOGUE_SCRIPTED_LINE_6);
-
-            pDialogueWindow->_41D08F_set_keyboard_control_group(num_menu_buttons, 1, 0, 2);
-            // TODO(Nik-RE-dev): initial number of buttons used only in non-simple houses now
-            //                   but this causes situation where dead character can talk to someone
-            //dword_F8B1E0 = pDialogueWindow->pNumPresenceButton;
-        }
-    }
-}
-
 //----- (00445C8B) --------------------------------------------------------
 int GetGreetType(signed int SpeakingNPC_ID) {
     if (SpeakingNPC_ID >= 0) {
