+++ src/GUI/GUIWindow.cpp
@@ -858,211 +858,6 @@ Color GetSkillColor(CharacterClassType uPlayerClass, CharacterSkillType uPlayerS
     return ui_character_skillinfo_cant_learn;
 }
 
-void ClickNPCTopic(DIALOGUE_TYPE topic) {
-    int pEventNumber;  // ecx@8
-    char *v12;         // eax@53
-    char *v13;         // eax@56
-    char *v14;         // eax@57
-    char *v15;         // eax@58
-    int pPrice;        // ecx@70
-
-    uDialogueType = (DIALOGUE_TYPE)(topic + 1);
-    NPCData *pCurrentNPCInfo = houseNpcs[currentHouseNpc].npc;
-    if (topic <= DIALOGUE_SCRIPTED_LINE_6) {
-        switch (topic) {
-        case DIALOGUE_13_hiring_related:
-            current_npc_text = BuildDialogueString(
-                pNPCStats->pProfessions[pCurrentNPCInfo->profession].pJoinText,
-                pParty->activeCharacterIndex() - 1, 0, HOUSE_INVALID, 0);
-            NPCHireableDialogPrepare();
-            dialogue_show_profession_details = false;
-            BackToHouseMenu();
-            return;
-        case DIALOGUE_SCRIPTED_LINE_1:
-            pEventNumber = pCurrentNPCInfo->dialogue_1_evt_id;
-            break;
-        case DIALOGUE_SCRIPTED_LINE_2:
-            pEventNumber = pCurrentNPCInfo->dialogue_2_evt_id;
-            break;
-        case DIALOGUE_SCRIPTED_LINE_3:
-            pEventNumber = pCurrentNPCInfo->dialogue_3_evt_id;
-            break;
-        case DIALOGUE_SCRIPTED_LINE_4:
-            pEventNumber = pCurrentNPCInfo->dialogue_4_evt_id;
-            break;
-        case DIALOGUE_SCRIPTED_LINE_5:
-            pEventNumber = pCurrentNPCInfo->dialogue_5_evt_id;
-            break;
-        case DIALOGUE_SCRIPTED_LINE_6:
-            pEventNumber = pCurrentNPCInfo->dialogue_6_evt_id;
-            break;
-        default:
-            BackToHouseMenu();
-            return;
-        }
-
-        if (pEventNumber < 200 || pEventNumber > 310) {
-            if (pEventNumber < 400 || pEventNumber > 410) {
-                if (pEventNumber == 139) {
-                    OracleDialogue();
-                } else {
-                    if (pEventNumber == 311) {
-                        // TODO(Nik-RE-dev): event 311 belongs to one of the teleports in Bracada
-                        __debugbreak();
-                        //openBountyHuntingDialogue();
-                    } else {
-                        current_npc_text.clear();
-                        activeLevelDecoration = (LevelDecoration *)1;
-                        eventProcessor(pEventNumber, 0, 1);
-                        activeLevelDecoration = nullptr;
-                    }
-                }
-            } else {
-                _dword_F8B1D8_last_npc_topic_menu = topic;
-                DrawJoinGuildWindow((GUILD_ID)(pEventNumber - 400));
-            }
-        } else {
-            _4B3FE5_training_dialogue(pEventNumber);
-        }
-        BackToHouseMenu();
-        return;
-    }
-    if (topic != DIALOGUE_HIRE_FIRE) {
-        if (topic == DIALOGUE_PROFESSION_DETAILS) {
-            // uBoxHeight = pCurrentNPCInfo->uProfession;
-            __debugbreak();  // probably hirelings found in buildings, not
-                             // present in MM7, changed
-                             // "pCurrentNPCInfo->uProfession - 1" to
-                             // "pCurrentNPCInfo->uProfession", have to check in
-                             // other versions whether it's ok
-            if (dialogue_show_profession_details) {
-                current_npc_text = BuildDialogueString(
-                    pNPCStats->pProfessions[pCurrentNPCInfo->profession].pJoinText,
-                    pParty->activeCharacterIndex() - 1, 0, HOUSE_INVALID, 0);
-            } else {
-                current_npc_text = BuildDialogueString(
-                    pNPCStats->pProfessions[pCurrentNPCInfo->profession].pBenefits,
-                    pParty->activeCharacterIndex() - 1, 0, HOUSE_INVALID, 0);
-            }
-            dialogue_show_profession_details = ~dialogue_show_profession_details;
-        } else {
-            if (topic == DIALOGUE_79_mastery_teacher) {
-                if (guild_membership_approved) {
-                    pParty->TakeGold(gold_transaction_amount);
-                    if (pParty->hasActiveCharacter()) {
-                        pParty->activeCharacter().SetSkillMastery(dword_F8B1AC_skill_being_taught, dword_F8B1B0_MasteryBeingTaught);
-                        pParty->activeCharacter().playReaction(SPEECH_SKILL_MASTERY_INC);
-                    }
-                    engine->_messageQueue->addMessageCurrentFrame(UIMSG_Escape, 1, 0);
-                }
-            } else {
-                if (topic == DIALOGUE_82_join_guild && guild_membership_approved) {
-                    // join guild
-                    pParty->TakeGold(gold_transaction_amount, true);
-                    for (Character &player : pParty->pCharacters)
-                        player.SetVariable(VAR_Award, dword_F8B1AC_award_bit_number);
-
-                    switch (_dword_F8B1D8_last_npc_topic_menu) {
-                    case DIALOGUE_SCRIPTED_LINE_1:
-                        pEventNumber = pCurrentNPCInfo->dialogue_1_evt_id;
-                        if (pEventNumber >= 400 && pEventNumber <= 416)
-                            pCurrentNPCInfo->dialogue_1_evt_id = 0;
-                        break;
-                    case DIALOGUE_SCRIPTED_LINE_2:
-                        pEventNumber = pCurrentNPCInfo->dialogue_2_evt_id;
-                        if (pEventNumber >= 400 && pEventNumber <= 416)
-                            pCurrentNPCInfo->dialogue_2_evt_id = 0;
-                        break;
-                    case DIALOGUE_SCRIPTED_LINE_3:
-                        pEventNumber = pCurrentNPCInfo->dialogue_3_evt_id;
-                        if (pEventNumber >= 400 && pEventNumber <= 416)
-                            pCurrentNPCInfo->dialogue_3_evt_id = 0;
-                        break;
-                    case DIALOGUE_SCRIPTED_LINE_4:
-                        pEventNumber = pCurrentNPCInfo->dialogue_4_evt_id;
-                        if (pEventNumber >= 400 && pEventNumber <= 416)
-                            pCurrentNPCInfo->dialogue_4_evt_id = 0;
-                        break;
-                    case DIALOGUE_SCRIPTED_LINE_5:
-                        pEventNumber = pCurrentNPCInfo->dialogue_5_evt_id;
-                        if (pEventNumber >= 400 && pEventNumber <= 416)
-                            pCurrentNPCInfo->dialogue_5_evt_id = 0;
-                        break;
-                    case DIALOGUE_SCRIPTED_LINE_6:
-                        pEventNumber = pCurrentNPCInfo->dialogue_6_evt_id;
-                        if (pEventNumber >= 400 && pEventNumber <= 416)
-                            pCurrentNPCInfo->dialogue_6_evt_id = 0;
-                        break;
-                    default:
-                        break;
-                    }
-                    engine->_messageQueue->addMessageCurrentFrame(UIMSG_Escape, 1, 0);
-                    if (pParty->hasActiveCharacter()) {
-                        pParty->activeCharacter().playReaction(SPEECH_JOINED_GUILD);
-                        BackToHouseMenu();
-                        return;
-                    }
-                }
-            }
-        }
-        BackToHouseMenu();
-        return;
-    }
-
-    if (!pParty->pHirelings[0].pName.empty() && !pParty->pHirelings[1].pName.empty()) {
-        GameUI_SetStatusBar(LSTR_HIRE_NO_ROOM);
-        BackToHouseMenu();
-        return;
-    }
-
-    if (pCurrentNPCInfo->profession != Burglar) {  // burglars have no hiring price
-        __debugbreak();  // probably hirelings found in buildings, not present
-                         // in MM7, changed "pCurrentNPCInfo->uProfession - 1"
-                         // to "pCurrentNPCInfo->uProfession", have to check in
-                         // other versions whether it's ok
-        pPrice =
-            pNPCStats->pProfessions[pCurrentNPCInfo->profession].uHirePrice;
-        if (pParty->GetGold() < (unsigned int)pPrice) {
-            GameUI_SetStatusBar(LSTR_NOT_ENOUGH_GOLD);
-            dialogue_show_profession_details = false;
-            uDialogueType = DIALOGUE_13_hiring_related;
-            current_npc_text = BuildDialogueString(
-                pNPCStats->pProfessions[pCurrentNPCInfo->profession].pJoinText,
-                pParty->activeCharacterIndex() - 1, 0, HOUSE_INVALID, 0);
-            if (pParty->hasActiveCharacter()) {
-                pParty->activeCharacter().playReaction(SPEECH_NOT_ENOUGH_GOLD);
-            }
-            GameUI_SetStatusBar(LSTR_NOT_ENOUGH_GOLD);
-            BackToHouseMenu();
-            return;
-        } else {
-            pParty->TakeGold(pPrice);
-        }
-    }
-
-    pCurrentNPCInfo->uFlags |= NPC_HIRED;
-    pParty->hirelingScrollPosition = 0;
-    pParty->CountHirelings();
-    if (!pParty->pHirelings[0].pName.empty()) {
-        pParty->pHirelings[1] = *pCurrentNPCInfo;
-        pParty->pHireling2Name = pCurrentNPCInfo->pName;
-    } else {
-        pParty->pHirelings[0] = *pCurrentNPCInfo;
-        pParty->pHireling1Name = pCurrentNPCInfo->pName;
-    }
-    pParty->hirelingScrollPosition = 0;
-    pParty->CountHirelings();
-    prepareHouse(window_SpeakInHouse->houseId());
-    dialog_menu_id = DIALOGUE_MAIN;
-
-    engine->_messageQueue->addMessageCurrentFrame(UIMSG_Escape, 1, 0);
-    if (pParty->hasActiveCharacter()) {
-        pParty->activeCharacter().playReaction(SPEECH_HIRE_NPC);
-    }
-
-    BackToHouseMenu();
-}
-
 std::string BuildDialogueString(const char *lpsz, uint8_t uPlayerID, ItemGen *a3, HOUSE_ID houseId, int a5, GameTime *a6) {
     std::string str = std::string(lpsz);
     return BuildDialogueString(str, uPlayerID, a3, houseId, a5, a6);
