--- src/GUI/UI/UIArena.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <array>
-
-#include "Utility/Geometry/Vec.h"
-
-extern std::array<Vec2i, 20> pMonsterArenaPlacements;
-
-void Arena_SelectionFightLevel();
-void ArenaFight();
