+++ src/GUI/UI/UIHouses.h
@@ -45,32 +45,12 @@ bool houseDialogPressEscape();
  */
 void playHouseSound(HOUSE_ID houseID, HouseSoundType type);
 
-/**
- * @offset 0x004B1ECE.
- *
- * @brief Oracle's 'I lost it!' dialog option
- */
-void OracleDialogue();
+void ClickNPCTopic(DIALOGUE_TYPE topic);
 
 /**
- * @offset 0x4B254D
+ * @offset 0x4B4224
  */
-std::string _4B254D_SkillMasteryTeacher(int trainerInfo);
-
-/**
- * @offset 0x4B3EF0
- */
-void DrawJoinGuildWindow(GUILD_ID guild_id);
-
-/**
- * @offset 0x4B29F2
- */
-const std::string &GetJoinGuildDialogueOption(GUILD_ID guild_id);
-
-/**
- * @offset 0x4B3FE5
- */
-void _4B3FE5_training_dialogue(int a4);
+void updateNPCTopics(int npc);
 
 /**
  * Type of NPC you can have dialogue with inside house.
@@ -153,6 +133,3 @@ extern IndexedArray<int, BUILDING_WEAPON_SHOP, BUILDING_DARK_GUILD> itemAmountIn
 
 extern std::vector<HouseNpcDesc> houseNpcs;
 extern int currentHouseNpc;
-
-extern AwardType dword_F8B1AC_award_bit_number;
-extern CharacterSkillType dword_F8B1AC_skill_being_taught;
